+++ Mantids30/Net_Sockets/socket_tls.h
@@ -499,7 +499,7 @@ public:
      * @param datalen data length in bytes
      * @return return the number of bytes read by the socket, zero for end of file and -1 for error.
      */
-    virtual int partialRead(void *data, const uint32_t &datalen) override;
+    virtual ssize_t partialRead(void *data, const uint32_t &datalen) override;
     /**
      * Write a data block to the TLS socket
      * note that this haves some limitations. some systems can only send 4k at time.
@@ -508,7 +508,7 @@ public:
      * @param datalen data length in bytes
      * @return return the number of bytes read by the socket, zero for end of file and -1 for error.
      */
-    virtual int partialWrite(const void *data, const uint32_t &datalen) override;
+    virtual ssize_t partialWrite(const void *data, const uint32_t &datalen) override;
 
     /////////////////////////
     // SSL functions:
